Support for timeout-based deconnnection/reconnection of ALSA device.
Direct hardware ALSA connection are always in blocking RW mode, which
poses problem when one wants to use other softwares with such
coonnections (for example MPD). The solution that this commit implements
allows to automatically close the ALSA device when no data is coming
through unicast (after a user-specified timeout), and immediately reopen
the same device when data is coming again (no timeout).

The case where an external soundcard/DAC is currently unavailable (for
example powered down or used by another app) is also taken into account.
In this case, the program will not fail but will retry the connection
until it can succeed.